Remove animation side effects
