+++ packages/dev/core/src/Actions/directActions.ts
@@ -6,6 +6,7 @@ import { Constants } from "../Engines/constants";
 import { RegisterClass } from "../Misc/typeStore";
 
 import type { ActionEvent } from "./actionEvent";
+import { BeginAnimation } from "core/Animations/animatable.core";
 
 /**
  * This defines an action responsible to toggle a boolean once triggered.
@@ -307,7 +308,7 @@ export class PlayAnimationAction extends Action {
      */
     public override execute(): void {
         const scene = this._actionManager.getScene();
-        scene.beginAnimation(this._target, this.from, this.to, this.loop);
+        BeginAnimation(scene, this._target, this.from, this.to, this.loop);
     }
 
     /**
