+++ packages/dev/core/src/Actions/interpolateValueAction.ts
@@ -7,6 +7,7 @@ import { Color3 } from "../Maths/math.color";
 import { Vector3, Matrix, Quaternion } from "../Maths/math.vector";
 import { Animation } from "../Animations/animation";
 import { RegisterClass } from "../Misc/typeStore";
+import { BeginDirectAnimation } from "core/Animations/animatable.core";
 
 /**
  * This defines an action responsible to change the value of a property
@@ -133,7 +134,7 @@ export class InterpolateValueAction extends Action {
             }
         };
 
-        scene.beginDirectAnimation(this._effectiveTarget, [animation], 0, 100, false, 1, wrapper);
+        BeginDirectAnimation(scene, this._effectiveTarget, [animation], 0, 100, false, 1, wrapper);
     }
 
     /**
