+++ packages/dev/core/src/Meshes/transformNode.ts
@@ -11,6 +11,7 @@ import type { Bone } from "../Bones/bone";
 import type { AbstractMesh } from "../Meshes/abstractMesh";
 import { Space } from "../Maths/math.axis";
 import { GetClass } from "../Misc/typeStore";
+import { BeginAnimation } from "core/Animations/animatable.core";
 
 const convertRHSToLHS = Matrix.Compose(Vector3.One(), Quaternion.FromEulerAngles(0, Math.PI, 0), Vector3.Zero());
 
@@ -1507,7 +1508,8 @@ export class TransformNode extends Node {
         }
 
         if (parsedTransformNode.autoAnimate) {
-            scene.beginAnimation(
+            BeginAnimation(
+                scene,
                 transformNode,
                 parsedTransformNode.autoAnimateFrom,
                 parsedTransformNode.autoAnimateTo,
