+++ packages/dev/core/src/Morph/morphTarget.ts
@@ -11,6 +11,7 @@ import { SerializationHelper } from "../Misc/decorators.serialization";
 import { GetClass } from "../Misc/typeStore";
 
 import type { Animation } from "../Animations/animation";
+import { BeginAnimation } from "core/Animations/animatable.core";
 
 /**
  * Defines a target to use with MorphTargetManager
@@ -313,7 +314,8 @@ export class MorphTarget implements IAnimatable {
             }
 
             if (serializationObject.autoAnimate && scene) {
-                scene.beginAnimation(
+                BeginAnimation(
+                    scene,
                     result,
                     serializationObject.autoAnimateFrom,
                     serializationObject.autoAnimateTo,
