+++ packages/dev/core/src/Particles/gpuParticleSystem.ts
@@ -52,6 +52,7 @@ import {
     CreatePointEmitter,
     CreateSphereEmitter,
 } from "./particleSystem.functions";
+import { BeginAnimation } from "core/Animations/animatable.core";
 
 /**
  * This represents a GPU particle system in Babylon
@@ -398,7 +399,7 @@ export class GPUParticleSystem extends BaseParticleSystem implements IDisposable
 
         // Animations
         if (this.beginAnimationOnStart && this.animations && this.animations.length > 0 && this._scene) {
-            this._scene.beginAnimation(this, this.beginAnimationFrom, this.beginAnimationTo, this.beginAnimationLoop);
+            BeginAnimation(this._scene, this, this.beginAnimationFrom, this.beginAnimationTo, this.beginAnimationLoop);
         }
     }
