+++ packages/dev/core/src/XR/features/WebXRNearInteraction.ts
@@ -26,6 +26,7 @@ import { QuadraticEase, EasingFunction } from "../../Animations/easing";
 // side effects
 import "../../Meshes/subMesh.project";
 import { Logger } from "core/Misc/logger";
+import { BeginDirectAnimation } from "core/Animations/animatable.core";
 
 type ControllerData = {
     xrController?: WebXRInputSource;
@@ -919,7 +920,7 @@ export class WebXRNearInteraction extends WebXRAbstractFeature {
 
         const touchCollisionMeshFunction = (isTouch: boolean) => {
             const action = isTouch ? touchAction : releaseAction;
-            meshCreationScene.beginDirectAnimation(touchCollisionMesh, [action], 0, 18, false, 1);
+            BeginDirectAnimation(meshCreationScene, touchCollisionMesh, [action], 0, 18, false, 1);
         };
 
         const hydrateCollisionMeshFunction = (isHydration: boolean) => {
@@ -927,7 +928,7 @@ export class WebXRNearInteraction extends WebXRAbstractFeature {
             if (isHydration) {
                 touchCollisionMesh.isVisible = true;
             }
-            meshCreationScene.beginDirectAnimation(touchCollisionMesh, [action], 0, 15, false, 1, () => {
+            BeginDirectAnimation(meshCreationScene, touchCollisionMesh, [action], 0, 15, false, 1, () => {
                 if (!isHydration) {
                     touchCollisionMesh.isVisible = false;
                 }
